Move converse-mam plugin into own folder
