+++ src/headless/headless.js
@@ -11,7 +11,7 @@ import "./plugins/chat/index.js";  // RFC-6121 Instant messaging
 import "./plugins/chatboxes.js";
 import "./plugins/disco.js";       // XEP-0030 Service discovery
 import "./plugins/headlines.js";   // Support for headline messages
-import "./plugins/mam.js";         // XEP-0313 Message Archive Management
+import "./plugins/mam/index.js";   // XEP-0313 Message Archive Management
 import "./plugins/muc/index.js";   // XEP-0045 Multi-user chat
 import "./plugins/ping.js";        // XEP-0199 XMPP Ping
 import "./plugins/pubsub.js";      // XEP-0060 Pubsub
